chore: sync with upstream akhilerm/tag-push-action v2.3.0
Bring product sources to upstream v2.3.0: entrypoint split (src/index.ts),
csv-parse and tooling updates, rebuilt dist. Keep StepSecurity
subscription (src/stepsecurity/subscription.ts), GHCR repo-copy image
pin, and package identity.
