+++ package.json
@@ -1,9 +1,9 @@
 {
   "name": "typescript-action",
-  "version": "2.2.0",
+  "version": "2.3.0",
   "private": true,
   "description": "TypeScript template action",
-  "main": "lib/main.js",
+  "main": "lib/index.js",
   "scripts": {
     "build": "tsc",
     "format": "prettier --write **/*.ts",
@@ -25,25 +25,29 @@
   "author": "step-security",
   "license": "MIT",
   "dependencies": {
-    "@actions/core": "^1.10.1",
-    "@actions/exec": "^1.0.4",
+    "@actions/core": "^1.11.1",
+    "@actions/exec": "^1.1.1",
     "axios": "^1.13.5",
-    "csv-parse": "^5.3.1"
+    "csv-parse": "^6.2.1"
   },
   "devDependencies": {
-    "@types/jest": "^27.0.1",
-    "@types/node": "^20.10.5",
+    "@types/jest": "^30.0.0",
+    "@types/node": "^25.9.2",
+    "@typescript-eslint/eslint-plugin": "^5.62.0",
     "@typescript-eslint/parser": "^5.62.0",
-    "@vercel/ncc": "^0.34.0",
+    "@vercel/ncc": "^0.44.0",
     "eslint": "^8.27.0",
-    "eslint-plugin-github": "^4.1.3",
-    "eslint-plugin-jest": "^27.6.0",
-    "eslint-plugin-prettier": "^5.1.2",
-    "jest": "^27.1.1",
-    "jest-circus": "^27.0.1",
-    "js-yaml": "^4.1.0",
-    "prettier": "3.1.0",
-    "ts-jest": "^27.0.5",
-    "typescript": "^4.2.4"
+    "eslint-plugin-github": "^4.10.2",
+    "eslint-plugin-jest": "^27.9.0",
+    "eslint-plugin-prettier": "^5.5.6",
+    "jest": "^30.4.2",
+    "jest-circus": "^30.4.2",
+    "js-yaml": "^4.2.0",
+    "prettier": "3.8.4",
+    "ts-jest": "^29.4.11",
+    "typescript": "^4.9.5"
+  },
+  "overrides": {
+    "undici": "^6.23.0"
   }
 }
