+++ src/index.ts
@@ -1,0 +1,7 @@
+/**
+ * The entrypoint for the action.
+ */
+
+import {run} from './main'
+
+run()
