+++ src/main.ts
@@ -2,57 +2,9 @@
 import {parse} from 'csv-parse/sync'
 import * as core from '@actions/core'
 import * as exec from '@actions/exec'
-import * as fs from 'fs'
-import axios, {isAxiosError} from 'axios'
+import {validateSubscription} from './stepsecurity/subscription'
 
-async function validateSubscription(): Promise<void> {
-  const eventPath = process.env.GITHUB_EVENT_PATH
-  let repoPrivate: boolean | undefined
-
-  if (eventPath && fs.existsSync(eventPath)) {
-    const eventData = JSON.parse(fs.readFileSync(eventPath, 'utf8'))
-    repoPrivate = eventData?.repository?.private
-  }
-
-  const upstream = 'akhilerm/tag-push-action'
-  const action = process.env.GITHUB_ACTION_REPOSITORY
-  const docsUrl =
-    'https://docs.stepsecurity.io/actions/stepsecurity-maintained-actions'
-
-  core.info('')
-  core.info('\u001b[1;36mStepSecurity Maintained Action\u001b[0m')
-  core.info(`Secure drop-in replacement for ${upstream}`)
-  if (repoPrivate === false)
-    core.info('\u001b[32m✓ Free for public repositories\u001b[0m')
-  core.info(`\u001b[36mLearn more:\u001b[0m ${docsUrl}`)
-  core.info('')
-
-  if (repoPrivate === false) return
-
-  const serverUrl = process.env.GITHUB_SERVER_URL || 'https://github.com'
-  const body: Record<string, string> = {action: action || ''}
-  if (serverUrl !== 'https://github.com') body.ghes_server = serverUrl
-  try {
-    await axios.post(
-      `https://agent.api.stepsecurity.io/v1/github/${process.env.GITHUB_REPOSITORY}/actions/maintained-actions-subscription`,
-      body,
-      {timeout: 3000}
-    )
-  } catch (error) {
-    if (isAxiosError(error) && error.response?.status === 403) {
-      core.error(
-        `\u001b[1;31mThis action requires a StepSecurity subscription for private repositories.\u001b[0m`
-      )
-      core.error(
-        `\u001b[31mLearn how to enable a subscription: ${docsUrl}\u001b[0m`
-      )
-      process.exit(1)
-    }
-    core.info('Timeout or API not reachable. Continuing to next step.')
-  }
-}
-
-async function run(): Promise<void> {
+export async function run(): Promise<void> {
   try {
     await validateSubscription()
     const source: string = core.getInput('src')
@@ -71,6 +23,7 @@ async function run(): Promise<void> {
       return
     }
 
+    // Use StepSecurity maintained repo-copy image (not upstream akhilerm/repo-copy)
     await exec.exec('docker', [
       'run',
       '--rm',
@@ -113,5 +66,3 @@ export async function getDestinationTags(): Promise<string[]> {
 
   return res.filter(item => item).map(pat => pat.trim())
 }
-
-run()
